+++ src/main/java/net/slipcor/mobstats/classes/NameHandler.java
@@ -13,12 +13,18 @@ public class NameHandler {
     public static OfflinePlayer findPlayer(String value) {
         OfflinePlayer result = null;
         for (OfflinePlayer off : Bukkit.getOfflinePlayers()) {
+            if (off == null || off.getName() == null) {
+                continue;
+            }
+
             if (off.getName().equalsIgnoreCase(value)) {
                 return off;
             }
+
             if (off.getPlayer() != null && off.getPlayer().getDisplayName().toLowerCase().contains(value.toLowerCase())) {
                 return off;
             }
+
             if (result == null && off.getName().toLowerCase().contains(value.toLowerCase())) {
                 result = off;
             }